Correct integer types

+++ lite/demo/cxx/lac_demo/lac.cc
@@ -24,7 +24,7 @@ LAC::LAC(const std::string &model_path,
          int threads,
          CODE_TYPE type)
     : _codetype(type),
-      _lod(std::vector<std::vector<size_t>>(1)),
+      _lod(std::vector<std::vector<uint64_t>>(1)),
       _word2id_dict(new std::unordered_map<std::string, int64_t>),
       _q2b_dict(new std::unordered_map<std::string, std::string>),
       _id2label_dict(new std::unordered_map<int64_t, std::string>) {
@@ -59,7 +59,7 @@ int LAC::feed_data(const std::vector<std::string> &querys) {
     shape += this->_seq_words.size();
     this->_lod[0].push_back(shape);
   }
-  this->_input_tensor->Resize({shape, 1});
+  this->_input_tensor->Resize({static_cast<int64_t>(shape), 1});
   this->_input_tensor->SetLoD(this->_lod);
   int64_t *input_d = this->_input_tensor->mutable_data<int64_t>();
   int index = 0;
